# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vector Technology Digital Services API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /wallet/recharge/order:
    post:
      operationId: createRechargeOrder
      tags: [wallet]
      summary: Create Razorpay order for recharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeInput"
      responses:
        "200":
          description: Razorpay order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RazorpayOrder"
  /wallet/recharge/verify:
    post:
      operationId: verifyRecharge
      tags: [wallet]
      summary: Verify and complete wallet recharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Recharge successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

  # SERVICES
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: List all active services
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ORDERS
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user's orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Purchase a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Insufficient balance or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # DASHBOARD SUMMARY (wow endpoint)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Wallet:
      type: object
      required: [balance, userId]
      properties:
        balance:
          type: integer
        userId:
          type: string
        totalRecharged:
          type: integer
        totalSpent:
          type: integer
    Transaction:
      type: object
      required: [id, userId, type, amount, description, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [success, failed, pending]
        referenceId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      required: [transactions, total, page, pages]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
    RechargeInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          minimum: 10
    RazorpayOrder:
      type: object
      required: [orderId, amount, currency, keyId]
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
    PaymentVerifyInput:
      type: object
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature]
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string

    # SERVICE SCHEMAS
    Service:
      type: object
      required: [id, title, description, pointsCost, deliveryTime, category, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        pointsCost:
          type: integer
        deliveryTime:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [active, inactive]
        features:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    ServiceListResponse:
      type: object
      required: [services, total, page, pages]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
...
    Order:
      type: object
      required: [id, userId, serviceId, serviceName, pointsCost, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        serviceCategory:
          type: ["string", "null"]
        pointsCost:
          type: integer
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        notes:
          type: ["string", "null"]
        deliveryTime:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [serviceId]
      properties:
        serviceId:
          type: string
        notes:
          type: string
...
    OrderListResponse:
      type: object
      required: [orders, total, page, pages]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer

    # PAYMENT SCHEMAS
...