# lib/api-zod/src/generated/types/razorpayOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vector Technology Digital Services API
 * OpenAPI spec version: 0.1.0
 */

export interface RazorpayOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}