# lib/api-zod/src/generated/types/rechargeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vector Technology Digital Services API
 * OpenAPI spec version: 0.1.0
 */

export interface RechargeInput {
  /** @minimum 10 */
  amount: number;
}